refactored a few things for readability, removed preview and listings sample as there was no way I could think of to make the preview work with current implementation
