+++ app/src/main/java/com/studhub/app/presentation/listing/browse/BrowseScreen.kt
@@ -1,78 +1,124 @@
 package com.studhub.app.presentation.listing.browse
 
 import BrowseContent
-import android.util.Log
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.studhub.app.R
-import com.studhub.app.annotations.ExcludeFromGeneratedTestCoverage
-import com.studhub.app.presentation.ui.common.input.RangeBar
-import com.studhub.app.presentation.ui.common.input.SearchBar
+import com.studhub.app.domain.model.Category
+import com.studhub.app.domain.model.Listing
+import com.studhub.app.domain.model.User
+import com.studhub.app.presentation.listing.browse.components.RangeBar
+import com.studhub.app.presentation.listing.browse.components.SearchBar
 import com.studhub.app.presentation.ui.common.text.BigLabel
 import kotlinx.coroutines.launch
 
 @Composable
 fun BrowseScreen(viewModel: BrowseViewModel = hiltViewModel(), navController: NavController) {
-    BigLabel(label = stringResource(R.string.listings_browsing_title))
     val scope = rememberCoroutineScope()
-    val search1 = remember {
-        mutableStateOf("")
-    }
-    val search2 = remember {
-        mutableStateOf("")
-    }
-    val search3 = remember {
-        mutableStateOf("")
-    }
     LaunchedEffect(viewModel) {
 
         /***
          *  If we want to use the fake listings,
          *  we can call viewModel.generateSampleListings()
          *  instead of viewModel.getAllListings()
-         *
          */
         scope.launch {
             viewModel.getCurrentListings()
         }
-
     }
+
     val listings = viewModel.listingsState.collectAsState().value
+    BrowseScreenListings(viewModel, navController, listings)
 
+}
+
+/**
+ * Main content of the screen where all the listings are displayed as well as the search fields
+ *
+ * @param viewModel takes a BrowseViewModel to be used to interact with the database
+ * @param navController needs to be given to allow navigation when clicking a specific listing
+ * @param listings the listings to be displayed
+ */
+@Composable
+fun BrowseScreenListings(
+    viewModel: BrowseViewModel = hiltViewModel(),
+    navController: NavController = rememberNavController(),
+    listings: List<Listing>
+) {
+    val search = remember {
+        mutableStateOf("")
+    }
+    val rangeMin = remember {
+        mutableStateOf("")
+    }
+    val rangeMax = remember {
+        mutableStateOf("")
+    }
+    BigLabel(label = stringResource(R.string.listings_browsing_title))
     Column {
-        SearchBar(search = search1, onSearch = {
-            viewModel.searchListings(search1.value)
+        SearchBar(search = search, onSearch = {
+            viewModel.searchListings(search.value)
         })
-        RangeBar("MIN....CHF", search = search2, onSearch = {
-            viewModel.rangeListings(search2.value, search3.value)
+        RangeBar("MIN....CHF", search = rangeMin, onSearch = {
+            viewModel.rangeListings(rangeMin.value, rangeMax.value)
         })
-        RangeBar("MAX....CHF",search = search3, onSearch = {
-            viewModel.rangeListings(search2.value, search3.value)
+        RangeBar("MAX....CHF", search = rangeMax, onSearch = {
+            viewModel.rangeListings(rangeMin.value, rangeMax.value)
         })
+        LoadListings(listings, navController)
+    }
+}
 
-        if (listings.isNotEmpty()) {
-            BrowseContent(listings, navController)
-        } else {
-            // Show a loading indicator while the listings are being fetched.
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator()
-                Spacer(modifier = Modifier.height(16.dp))
-            }
-
+@Composable
+fun LoadListings(listings: List<Listing>, navController: NavController) {
+    if (listings.isNotEmpty()) {
+        BrowseContent(listings, navController)
+    } else {
+        // Show a loading indicator while the listings are being fetched.
+        Box(
+            modifier = Modifier.fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            CircularProgressIndicator()
+            Spacer(modifier = Modifier.height(16.dp))
         }
     }
+}
 
+fun fakeListings(): List<Listing> {
+    return listOf(
+        Listing(
+            id = "33",
+            name = "Algebra for the dummies",
+            seller = User(firstName = "Jacky", lastName = "Chan"),
+            categories = listOf(Category(name = "Books")),
+            description = "Really great book to learn Algebra for entry level readers.",
+            price = 34.50F
+        ),
+        Listing(
+            id = "32",
+            name = "Brand new Nike Air One",
+            seller = User(firstName = "Kristina", lastName = "Gordova"),
+            categories = listOf(Category(name = "Clothing")),
+            description = "Brand new shoes, full white and ready for any custom work if needed.",
+            price = 194.25F
+        ),
+        Listing(
+            id = "31",
+            name = "Super VTT 2000 with custom paint",
+            seller = User(firstName = "Marc", lastName = "Marquez"),
+            categories = listOf(Category(name = "Mobility")),
+            description = "Robust bike for downhill riding and will " +
+                    "look absolutely unique with this custom paint work",
+            price = 1500F
+        )
+    )
 }
